refactor(adapters): rename to legacy/v2-5/ + drop unused imports
Two review-driven refinements to the adapter registry:

1. Rename `src/lib/adapters/v3-to-v2-5/` → `src/lib/adapters/legacy/v2-5/`.
   Per ad-tech-product review: real ad-tech compat layers carry N=1
   active legacy shim with a deprecation runway (OpenRTB, Prebid, GAM
   all behave this way). Encoding `<sdk-version>-to-<seller-version>/`
   commits the codebase to a matrix nobody will staff. `legacy/v2-5/`
   reflects "exceptional, time-boxed compat" — when the SDK pin moves
   from v3 to v4, this directory continues to hold the v2.5 shim and a
   sibling `legacy/v3/` joins for v3 sellers.

   getV3ToV25Adapter → getV25Adapter (sdk-version implicit)
   listV3ToV25AdapterTools → listV25AdapterTools

2. Drop `SingleAgentClient.ts` imports that became unused after the
   dispatch refactor (adaptCreateMediaBuyRequestForV2,
   adaptUpdateMediaBuyRequestForV2, normalizeMediaBuyResponse,
   adaptSyncCreativesRequestForV2, normalizeFormatsResponse,
   normalizePreviewCreativeResponse, normalizeGetProductsResponse,
   adaptGetProductsRequestForV2). Code-reviewer flagged as a
   noUnusedLocals failure waiting to happen. The helpers are still
   exported from utils/*-adapter.ts for downstream callers using them
   directly — only the dispatch-site imports go away.

5565 lib tests pass / 7 pre-existing skipped / 0 fail.
