+++ src/lib/adapters/legacy/v2-5/create_media_buy.ts
@@ -1,5 +1,5 @@
-import { adaptCreateMediaBuyRequestForV2, normalizeMediaBuyResponse } from '../../utils/creative-adapter';
-import type { CreateMediaBuyRequest, CreateMediaBuyResponse } from '../../types/v2-5';
+import { adaptCreateMediaBuyRequestForV2, normalizeMediaBuyResponse } from '../../../utils/creative-adapter';
+import type { CreateMediaBuyRequest, CreateMediaBuyResponse } from '../../../types/v2-5';
 import type { AdapterPair } from './types';
 
 export const createMediaBuyAdapter: AdapterPair<unknown, CreateMediaBuyRequest, CreateMediaBuyResponse, unknown> = {
