+++ src/lib/adapters/legacy/v2-5/get_products.ts
@@ -1,5 +1,5 @@
-import { adaptGetProductsRequestForV2, normalizeGetProductsResponse } from '../../utils/pricing-adapter';
-import type { GetProductsRequest, GetProductsResponse } from '../../types/v2-5';
+import { adaptGetProductsRequestForV2, normalizeGetProductsResponse } from '../../../utils/pricing-adapter';
+import type { GetProductsRequest, GetProductsResponse } from '../../../types/v2-5';
 import type { AdapterPair } from './types';
 
 /**
