+++ src/lib/adapters/legacy/v2-5/types.ts
