+++ test/lib/legacy-v2-5-adapter-registry.test.js
@@ -15,7 +15,7 @@
 const { test, describe } = require('node:test');
 const assert = require('node:assert');
 
-const { getV3ToV25Adapter, listV3ToV25AdapterTools } = require('../../dist/lib/adapters/v3-to-v2-5');
+const { getV25Adapter, listV25AdapterTools } = require('../../dist/lib/adapters/legacy/v2-5');
 const { adaptGetProductsRequestForV2, normalizeGetProductsResponse } = require('../../dist/lib/utils/pricing-adapter');
 const {
   adaptCreateMediaBuyRequestForV2,
@@ -36,21 +36,21 @@ describe('v3 → v2.5 adapter registry', () => {
       'list_creative_formats',
       'preview_creative',
     ]);
-    const registered = new Set(listV3ToV25AdapterTools());
+    const registered = new Set(listV25AdapterTools());
     for (const tool of expected) {
       assert.ok(registered.has(tool), `${tool} must be registered`);
     }
   });
 
-  test('getV3ToV25Adapter returns undefined for unregistered tools', () => {
-    assert.strictEqual(getV3ToV25Adapter('not_a_real_tool'), undefined);
+  test('getV25Adapter returns undefined for unregistered tools', () => {
+    assert.strictEqual(getV25Adapter('not_a_real_tool'), undefined);
   });
 
   describe('request-side adapters route to the existing utils/* helpers', () => {
     test('get_products', () => {
       const input = { brief: 'test', buying_mode: 'brief', brand: { domain: 'example.com' } };
       const direct = adaptGetProductsRequestForV2(structuredClone(input));
-      const viaRegistry = getV3ToV25Adapter('get_products').adaptRequest(structuredClone(input));
+      const viaRegistry = getV25Adapter('get_products').adaptRequest(structuredClone(input));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
@@ -64,14 +64,14 @@ describe('v3 → v2.5 adapter registry', () => {
         idempotency_key: '11111111-1111-1111-1111-111111111111',
       };
       const direct = adaptCreateMediaBuyRequestForV2(structuredClone(input));
-      const viaRegistry = getV3ToV25Adapter('create_media_buy').adaptRequest(structuredClone(input));
+      const viaRegistry = getV25Adapter('create_media_buy').adaptRequest(structuredClone(input));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
     test('update_media_buy', () => {
       const input = { media_buy_id: 'mb', idempotency_key: '22222222-2222-2222-2222-222222222222' };
       const direct = adaptUpdateMediaBuyRequestForV2(structuredClone(input));
-      const viaRegistry = getV3ToV25Adapter('update_media_buy').adaptRequest(structuredClone(input));
+      const viaRegistry = getV25Adapter('update_media_buy').adaptRequest(structuredClone(input));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
@@ -89,16 +89,16 @@ describe('v3 → v2.5 adapter registry', () => {
         idempotency_key: '33333333-3333-3333-3333-333333333333',
       };
       const direct = adaptSyncCreativesRequestForV2(structuredClone(input));
-      const viaRegistry = getV3ToV25Adapter('sync_creatives').adaptRequest(structuredClone(input));
+      const viaRegistry = getV25Adapter('sync_creatives').adaptRequest(structuredClone(input));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
     test('list_creative_formats and preview_creative are pass-through on request side', () => {
       const formatsInput = { type: 'video' };
-      assert.deepStrictEqual(getV3ToV25Adapter('list_creative_formats').adaptRequest(formatsInput), formatsInput);
+      assert.deepStrictEqual(getV25Adapter('list_creative_formats').adaptRequest(formatsInput), formatsInput);
 
       const previewInput = { creative_id: 'c' };
-      assert.deepStrictEqual(getV3ToV25Adapter('preview_creative').adaptRequest(previewInput), previewInput);
+      assert.deepStrictEqual(getV25Adapter('preview_creative').adaptRequest(previewInput), previewInput);
     });
   });
 
@@ -106,40 +106,40 @@ describe('v3 → v2.5 adapter registry', () => {
     test('get_products', () => {
       const v25Resp = { products: [{ product_id: 'p', name: 'P' }] };
       const direct = normalizeGetProductsResponse(structuredClone(v25Resp));
-      const viaRegistry = getV3ToV25Adapter('get_products').normalizeResponse(structuredClone(v25Resp));
+      const viaRegistry = getV25Adapter('get_products').normalizeResponse(structuredClone(v25Resp));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
     test('create_media_buy', () => {
       const v25Resp = { media_buy_id: 'mb', status: 'completed' };
       const direct = normalizeMediaBuyResponse(structuredClone(v25Resp));
-      const viaRegistry = getV3ToV25Adapter('create_media_buy').normalizeResponse(structuredClone(v25Resp));
+      const viaRegistry = getV25Adapter('create_media_buy').normalizeResponse(structuredClone(v25Resp));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
     test('update_media_buy uses the same media-buy normalizer', () => {
       const v25Resp = { media_buy_id: 'mb', status: 'completed' };
       const direct = normalizeMediaBuyResponse(structuredClone(v25Resp));
-      const viaRegistry = getV3ToV25Adapter('update_media_buy').normalizeResponse(structuredClone(v25Resp));
+      const viaRegistry = getV25Adapter('update_media_buy').normalizeResponse(structuredClone(v25Resp));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
     test('list_creative_formats', () => {
       const v25Resp = { formats: [] };
       const direct = normalizeFormatsResponse(structuredClone(v25Resp));
-      const viaRegistry = getV3ToV25Adapter('list_creative_formats').normalizeResponse(structuredClone(v25Resp));
+      const viaRegistry = getV25Adapter('list_creative_formats').normalizeResponse(structuredClone(v25Resp));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
     test('preview_creative', () => {
       const v25Resp = { renders: [] };
       const direct = normalizePreviewCreativeResponse(structuredClone(v25Resp));
-      const viaRegistry = getV3ToV25Adapter('preview_creative').normalizeResponse(structuredClone(v25Resp));
+      const viaRegistry = getV25Adapter('preview_creative').normalizeResponse(structuredClone(v25Resp));
       assert.deepStrictEqual(viaRegistry, direct);
     });
 
     test('sync_creatives has no response normalizer (pass-through)', () => {
-      const pair = getV3ToV25Adapter('sync_creatives');
+      const pair = getV25Adapter('sync_creatives');
       assert.strictEqual(pair.normalizeResponse, undefined);
     });
   });
